requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-core==1.8.1
azure-common==1.1.25
azure-nspkg==3.0.2
azure-functions==1.3.1
azure-storage-common==2.1.0
azure-storage-nspkg==3.1.0
azure-storage-blob==12.5.0
azure-cosmos==4.1.0
PyYAML==5.3.1
jinja2==2.11.2
flask==1.1.2
plotly==4.10.0
kaleido==0.0.3.post1
pandas==1.0.5
uk-covid19==1.2.0
flask-minify==0.27
werkzeug==1.0.1
redis==3.5.3
pytz==2020.01
flask-caching==1.9.0
python-dotenv==0.14.0
tqdm==4.51.0